update: Article, Author, Comment, Tags 타입 작성
