+++ types/Article.ts
@@ -1,3 +1,5 @@
+import { Author } from "@/types/Author";
+
 export interface Articles {
   articles: Article[];
   articlesCount: number;
@@ -14,11 +16,4 @@ export interface Article {
   favoritessCount: number;
   createdAt: string;
   updatedAt: string;
-}
-
-export interface Author {
-  username: string;
-  bio: string;
-  image: string;
-  following: boolean;
 }
