+++ types/Author.ts
@@ -1,0 +1,6 @@
+export interface Author {
+  username: string;
+  bio: string;
+  image: string;
+  following: boolean;
+}
