+++ types/Tags.ts
@@ -1,0 +1,3 @@
+export interface Tags {
+  tags: string[];
+}
